refactor: `NewPutClient` is a `testutil`
