+++ pkg/client/spaceblobadd_test.go
@@ -2,9 +2,6 @@ package client_test
 
 import (
 	"bytes"
-	"fmt"
-	"io"
-	"net/http"
 	"testing"
 
 	"github.com/storacha/go-ucanto/core/delegation"
@@ -15,31 +12,11 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// blobPutTransport is an [http.RoundTripper] (an [http.Client] transport) that
-// accepts blob PUTs and remembers what was received.
-type blobPutTransport struct {
-	receivedBlobs [][]byte
-}
-
-var _ http.RoundTripper = (*blobPutTransport)(nil)
-
-func (r *blobPutTransport) RoundTrip(req *http.Request) (*http.Response, error) {
-	blob, err := io.ReadAll(req.Body)
-	if err != nil {
-		return nil, fmt.Errorf("reading blob from request: %w", err)
-	}
-	r.receivedBlobs = append(r.receivedBlobs, blob)
-
-	return &http.Response{
-		StatusCode: 200,
-	}, nil
-}
-
 func TestSpaceBlobAdd(t *testing.T) {
 	space, err := ed25519signer.Generate()
 	require.NoError(t, err)
 
-	blobPutTransport := blobPutTransport{}
+	putClient := testutil.NewPutClient()
 
 	c, err := testutil.SpaceBlobAddClient()
 	require.NoError(t, err)
@@ -53,10 +30,8 @@ func TestSpaceBlobAdd(t *testing.T) {
 
 	testBlob := bytes.NewReader([]byte("test"))
 
-	_, _, err = c.SpaceBlobAdd(testContext(t), testBlob, space.DID(), client.WithPutClient(&http.Client{
-		Transport: &blobPutTransport,
-	}))
+	_, _, err = c.SpaceBlobAdd(testContext(t), testBlob, space.DID(), client.WithPutClient(putClient))
 	require.NoError(t, err)
 
-	require.ElementsMatch(t, [][]byte{[]byte("test")}, blobPutTransport.receivedBlobs)
+	require.ElementsMatch(t, [][]byte{[]byte("test")}, testutil.ReceivedBlobs(putClient))
 }
